added comments

+++ menu_element.cpp
@@ -60,7 +60,7 @@ MenuElement::MenuElement(const char* text, gef::Vector4 pos, float scale, float 
 
 }
 
-void MenuElement::Update()
+void MenuElement::Update() // if selected, increase size and alpha
 {
 	if (selected_) {
 		updateSprite(1.1f, 0.8f);
@@ -97,8 +97,9 @@ void MenuElement::updateSprite(float scale, float alpha)
 
 void MenuElement::Render()
 {
-	sprite_renderer_->DrawSprite(sprite_);
+	sprite_renderer_->DrawSprite(sprite_); // draw "button"
 
+	// render button text
 	font_->RenderText(
 		sprite_renderer_,
 		text_pos_,
@@ -107,21 +108,21 @@ void MenuElement::Render()
 		gef::TJ_CENTRE,
 		text_);
 
-	if (type_ == TOGGLE) 
+	if (type_ == TOGGLE)  // if toggle button, render state based on toggle_ pointer
 	{
 		font_->RenderText(
 			sprite_renderer_,
-			text_pos_ + gef::Vector4((int)strlen(text_) * 16, 0.f, 0.f),
+			text_pos_ + gef::Vector4(float((int)strlen(text_) * 16), 0.f, 0.f),
 			text_scale_,
-			*toggle_ ? 0xff00ff00 : 0xff0000ff,
+			*toggle_ ? 0xff00ff00 : 0xff0000ff, // if true, green, else red
 			gef::TJ_CENTRE,
-			"[%s]", *toggle_ ? "ON" : "OFF");
+			"[%s]", *toggle_ ? "ON" : "OFF"); // if true, "ON", else "OFF"
 	}
 	else if (type_ == SLIDER) 
 	{
 		font_->RenderText(
 			sprite_renderer_,
-			text_pos_ + gef::Vector4((int)strlen(text_)*11, 0.f, 0.f),
+			text_pos_ + gef::Vector4(float((int)strlen(text_) * 11), 0.f, 0.f),
 			text_scale_,
 			0xff00ff00,
 			gef::TJ_CENTRE,
